Fix high RAM usage
Fixed ram usage bug, by dynamically determining relevant current-time points to plot. This is done by only plotting points that go over-under a dynamic average calculated, or by date. In this case, points with currents over 115% or under 85% of the average will be plotted (and the average reinitialized to this outlier point), or if at least 1800 seconds have passed since the last plotted point.

The ram usage was reduced over 95% with this fix.